pyproject.toml
[project]
name = "yt-fastapi-beginner-fast-track-graphql"
version = "0.1.0"
description = "A tutorial project demonstrating how to build a GraphQL API with FastAPI, SQLAlchemy, and PostgreSQL for a simple blog system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.16.4",
    "black==25.1.0",
    "fastapi==0.116.1",
    "graphene==3.4.3",
    "graphene-sqlalchemy==3.0.0rc2",
    "httptools==0.6.4",
    "psycopg==3.2.9",
    "psycopg-binary==3.2.9",
    "python-dotenv==1.1.1",
    "sqlalchemy==2.0.41",
    "starlette-graphene3==0.6.0",
    "uvicorn==0.35.0",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
    # Testing dependencies
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.24.0",
    "factory-boy>=3.2.0",
]
